+++ packages/nextjs/tsup.config.ts
@@ -3,7 +3,6 @@ import type { Options } from 'tsup';
 export const tsup: Options = {
   dts: true,
   entryPoints: ['src/index.ts', 'src/middleware/index.ts'],
-  // `aws-amplify` is external, but sub-dependencies weren't automatically externalized ("require" statements were included)
   external: ['next', 'react', /^@supabase\//],
   format: ['cjs'],
   //   inject: ['src/react-shim.js'],