+++ packages/sveltekit/src/handlers/user.ts
@@ -6,11 +6,15 @@ import {
   parseCookie,
   jwtDecoder,
   TOKEN_REFRESH_MARGIN,
-  type User
+  CookieNotParsed,
+  JWTPayloadFailed,
+  AuthHelperError,
+  AccessTokenNotFound,
+  CookieNotFound
 } from '@supabase/auth-helpers-shared';
 import { skHelper } from '../instance';
-import getUser from '../utils/getUser';
-import type { Locals } from '../types';
+import { getUser, saveTokens } from '../utils/getUser';
+import logger from '../utils/log';
 
 export interface HandleUserOptions {
   cookieOptions?: CookieOptions;
@@ -24,32 +28,43 @@ export const handleUser = (options: HandleUserOptions = {}) => {
 
     try {
       if (!req.headers.has('cookie')) {
-        throw new Error('Not able to parse cookies');
+        throw new CookieNotFound();
       }
 
       const cookieOptions = { ...COOKIE_OPTIONS, ...options.cookieOptions };
-      const tokenRefreshMargin = options.tokenRefreshMargin ?? TOKEN_REFRESH_MARGIN;
+      const tokenRefreshMargin =
+        options.tokenRefreshMargin ?? TOKEN_REFRESH_MARGIN;
 
       const cookies = parseCookie(req.headers.get('cookie'));
+      if (!cookies) {
+        throw new CookieNotParsed();
+      }
       const access_token = cookies[`${cookieOptions.name}-access-token`];
 
       if (!access_token) {
-        throw new Error('No cookie found!');
+        throw new AccessTokenNotFound();
       }
 
       // Get payload from cached access token.
       const jwtUser = jwtDecoder(access_token);
       if (!jwtUser?.exp) {
-        throw new Error('Not able to parse JWT payload!');
+        throw new JWTPayloadFailed();
       }
       const timeNow = Math.round(Date.now() / 1000);
       if (jwtUser.exp < timeNow + tokenRefreshMargin) {
-        const res = await resolve(event);
         // JWT is expired, let's refresh from Gotrue
-        const response = await getUser({ req, res }, { cookieOptions, tokenRefreshMargin });
-        event.locals.user = response.user;
-        event.locals.accessToken = response.accessToken;
-        return await resolve(event);
+        const session = await getUser(req, {
+          cookieOptions,
+          tokenRefreshMargin
+        });
+        event.locals.user = session.user;
+        event.locals.accessToken = session.accessToken;
+        const res = await resolve(event);
+        await saveTokens({ req, res }, session, {
+          cookieOptions,
+          tokenRefreshMargin
+        });
+        return res;
       } else {
         // Transform JWT and add note that it is cached from JWT.
         const user = {
@@ -71,9 +86,20 @@ export const handleUser = (options: HandleUserOptions = {}) => {
         return await resolve(event);
       }
     } catch (e) {
-      const error = e as ApiError;
+      if (e instanceof JWTPayloadFailed) {
+        logger.info('JWTPayloadFailed error has happened!');
+        event.locals.error = e.toObj();
+      } else if (e instanceof CookieNotFound) {
+        logger.warn(e.toString());
+      } else if (e instanceof AuthHelperError) {
+        logger.info('AuthHelperError error has happened!');
+        logger.error(e.toString());
+      } else {
+        const error = e as ApiError;
+        logger.error(error.message);
+      }
+
       event.locals.user = null;
-      event.locals.error = error.message;
       event.locals.accessToken = null;
       return await resolve(event);
     }