+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @supabase/auth-helpers-nextjs
 
+## 0.2.3
+
+### Patch Changes
+
+- 69fefcb: Change logger to be a wrapper for console
+
+## 0.2.2
+
+### Patch Changes
+
+- 38ccf1c: Logger can be imported in your own project for setting log levels
+- Updated dependencies [38ccf1c]
+  - @supabase/auth-helpers-shared@0.1.2
+
+## 0.2.1
+
+### Patch Changes
+
+- 9dda264: Add better error handling and error codes
+- Updated dependencies [9dda264]
+  - @supabase/auth-helpers-shared@0.1.1
+
 ## 0.2.0
 
 ### Minor Changes