+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # shared
 
+## 0.1.2
+
+### Patch Changes
+
+- 38ccf1c: Add new to string method to the base error class
+
+## 0.1.1
+
+### Patch Changes
+
+- 9dda264: Add better error handling and error codes
+
 ## 0.1.0
 
 ### Minor Changes