+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @supabase/auth-helpers-svelte
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [38ccf1c]
+  - @supabase/auth-helpers-shared@0.1.2
+
+## 0.4.3
+
+### Patch Changes
+
+- 9dda264: Add better error handling and error codes
+- 9dda264: Fix no cookie found issue
+- Updated dependencies [9dda264]
+  - @supabase/auth-helpers-shared@0.1.1
+
 ## 0.4.2
 
 ### Patch Changes